+++ core/parameters/src/vm.rs
@@ -175,6 +175,9 @@ pub struct Config {
     /// Enable the `EthImplicitAccounts` protocol feature.
     pub eth_implicit_accounts: bool,
 
+    /// Enable using global contract for ETH implicit accounts.
+    pub eth_implicit_global_contract: bool,
+
     /// Whether to discard custom sections.
     pub discard_custom_sections: bool,
 
@@ -213,8 +216,10 @@ impl Config {
         self.limit_config.max_gas_burnt = Gas::MAX;
     }
 
+    /// Enable all protocol features. Only used for gas cost estimations.
     pub fn enable_all_features(&mut self) {
         self.eth_implicit_accounts = true;
+        self.eth_implicit_global_contract = true;
         self.global_contract_host_fns = true;
     }
 }