+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * The contract runtime has been upgraded to use the new Wasmtime-based runtime;
 * The contract runtime now allows for bulk memory instructions in Wasm code.
 * Yield/Resume improvements - from now on, calling `promise_yield_resume` after `promise_yield_create` will always work. Previously there were some cases where calling `resume` would fail if called too quickly after `create`. See the [github issue](https://github.com/near/nearcore/issues/14904) for details. Strictly speaking, this could be a breaking change for contracts which expected the resume to fail in these situations, but it shouldn't affect normal use cases.
+* A promise with a single `DeleteAccount` action and no input promises will be executed immediately after the receipt that created it, before any other receipts for this account.
+* ETH implicit accounts now use a globally deployed contract instead of a contract embedded in the neard binary.
 
 ### Non-protocol Changes
 * Replace polling for transaction status with event notifications to improve jsonrpc response latency